requirements.txt
scrapy==2.5.1
itemadapter==0.4.0
instaloader==4.8.1
black==21.12b0
webdriver-manager==3.5.2
selenium==4.1.0
parsel==1.6.0
python-decouple==3.6
